Update Script_2_IRC_Calculation.py
Corrected so that the Bcomparer function now correctly compares species that converged in terms of energy in a one-on-one fashion and not all converged species together.

e.g. conformer 1 and 2 converged & 3 and 4 converged.

Before the program would compare 1 vs. 2, 1 vs. 3 and 1 vs. 4 and analogously for 2 and 3, which resulted in some duplicates being saved. Now it correctly compares 1 vs. 2 and 3 vs. 4 only.